+++ quickPatterns.cpp
@@ -25,12 +25,14 @@ void quickPatterns::draw() {
     this->lightStrand->clearMain();
 
     this->currentScene->draw(this->lightStrand->leds, this->lightStrand->numLeds);
+    //TODO: double write fixes Teensy 3.2, investigate
+//    this->currentScene->draw(this->lightStrand->leds, this->lightStrand->numLeds);
   }
 
 }
 
 
-// Quick add
+// Quick add pattern
 
 qpPattern &quickPatterns::addPattern(qpPattern *pattern) {
 
@@ -39,7 +41,7 @@ qpPattern &quickPatterns::addPattern(qpPattern *pattern) {
 
 
 // Returns requested scene or automatically adds and returns a new scene if index does not yet exist
-qpScene &quickPatterns::scene(int index) {
+qpScene &quickPatterns::scene(byte index) {
 
   if(index > (this->scenes.numElements - 1))
     return this->newScene();
@@ -63,7 +65,7 @@ qpScene &quickPatterns::newScene() {
 
 // ~ Access
 
-qpLayer &quickPatterns::layer(int layerIndex) {
+qpLayer &quickPatterns::layer(byte layerIndex) {
 
   return this->scene(0).layer(layerIndex);
 }
@@ -80,7 +82,7 @@ void quickPatterns::nextScene() {
 }
 
 
-void quickPatterns::playScene(int index) {
+void quickPatterns::playScene(byte index) {
 
   if(index >= this->scenes.numElements)
     return;
@@ -95,7 +97,7 @@ void quickPatterns::playScene(int index) {
 
 void quickPatterns::playRandomScene() {
 
-  int index;
+  byte index;
   do {
     index = random8(0, (this->scenes.numElements));
   } while(index == this->sceneIndex);
@@ -106,17 +108,17 @@ void quickPatterns::playRandomScene() {
 
 // Quick access operators
 
-qpPattern&quickPatterns::operator()(int layerIndex) {
+qpPattern&quickPatterns::operator()(byte layerIndex) {
 
   return this->scene(0).layer(layerIndex).pattern(0);
 }
 
-qpPattern&quickPatterns::operator()(int sceneIndex, int layerIndex) {
+qpPattern&quickPatterns::operator()(byte sceneIndex, byte layerIndex) {
 
   return this->scene(sceneIndex).layer(layerIndex).pattern(0);
 }
 
-qpPattern&quickPatterns::operator()(int sceneIndex, int layerIndex, int patternIndex) {
+qpPattern&quickPatterns::operator()(byte sceneIndex, byte layerIndex, byte patternIndex) {
 
   return this->scene(sceneIndex).layer(layerIndex).pattern(patternIndex);
 }
