+++ quickPatterns.h
@@ -4,6 +4,7 @@
 #define DIR_FORWARD 1
 #define DIR_REVERSE -1
 
+#include <Arduino.h>
 #include <qpLinkedList.h>
 #include <qpPattern.h>
 #include <qpLayer.h>
@@ -43,24 +44,24 @@ class quickPatterns {
 
     // Access
 
-    qpPattern &pattern(int patternIndex); //returns specified patter on layer 0
+    qpPattern &pattern(byte patternIndex); //returns specified patter on layer 0
     qpPattern &addPattern(qpPattern *pattern); //creates a new layer and adds passed pattern as pattern 0
     qpPattern &samePattern() { return this->sameScene().sameLayer().samePattern(); }
 
-    qpLayer &layer(int layerIndex);
+    qpLayer &layer(byte layerIndex);
     qpLayer &sameLayer() { return this->sameScene().sameLayer(); }
 
-    qpScene &scene(int sceneIndex);
+    qpScene &scene(byte sceneIndex);
     qpScene &newScene();
     qpScene &sameScene() { return *this->lastReferencedScene; }
 
-    qpPattern &operator()(int layerIndex); //returns pattern 0 from the specified layer in scene 0
-    qpPattern &operator()(int sceneIndex, int layerIndex); //returns pattern 0 from the specified layer in the specified scene
-    qpPattern &operator()(int sceneIndex, int layerIndex, int patternIndex);
+    qpPattern &operator()(byte layerIndex); //returns pattern 0 from the specified layer in scene 0
+    qpPattern &operator()(byte sceneIndex, byte layerIndex); //returns pattern 0 from the specified layer in the specified scene
+    qpPattern &operator()(byte sceneIndex, byte layerIndex, byte patternIndex);
 
     // Scene navigation
 
-    void playScene(int index);
+    void playScene(byte index);
     void nextScene();
     void playRandomScene();
